# Remove SortKeyCursor (~5-10% faster merge)