# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Part of https://github.com/apache/arrow-datafusion/issues/5882
Relates to https://github.com/apache/arrow-datafusion/issues/5879

# Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

Replaces SortKeyCursor with a simplified, crate-private version. I couldn't see an obvious reason why SortKeyCursor was public other than to enable integration tests. Especially now that arrow-row exists, I don't see a compelling reason to keep it around.

```
merge sorted i64        time:   [8.7181 ms 8.7273 ms 8.7378 ms]
                        change: [-8.6612% -8.5375% -8.3983%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

sort merge i64          time:   [8.9273 ms 8.9396 ms 8.9521 ms]
                        change: [-8.4047% -8.2575% -8.0995%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

merge sorted f64        time:   [8.6883 ms 8.7002 ms 8.7140 ms]
                        change: [-8.4371% -8.2785% -8.1206%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

sort merge f64          time:   [8.9428 ms 8.9550 ms 8.9680 ms]
                        change: [-7.8564% -7.7071% -7.5387%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

merge sorted utf8 low cardinality
                        time:   [6.2272 ms 6.2357 ms 6.2448 ms]
                        change: [-14.595% -14.461% -14.314%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

sort merge utf8 low cardinality
                        time:   [6.8843 ms 6.9109 ms 6.9380 ms]
                        change: [-13.368% -12.957% -12.521%] (p = 0.00 < 0.05)
                        Performance has improved.

merge sorted utf8 high cardinality
                        time:   [10.147 ms 10.224 ms 10.307 ms]
                        change: [-6.2273% -5.3649% -4.5217%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 21 outliers among 100 measurements (21.00%)
  3 (3.00%) high mild
  18 (18.00%) high severe

sort merge utf8 high cardinality
                        time:   [11.191 ms 11.254 ms 11.321 ms]
                        change: [-7.4775% -6.7869% -6.0794%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

merge sorted utf8 tuple time:   [17.150 ms 17.169 ms 17.188 ms]
                        change: [-4.9041% -4.7286% -4.5686%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

sort merge utf8 tuple   time:   [20.286 ms 20.399 ms 20.511 ms]
                        change: [-5.5443% -4.8463% -4.1378%] (p = 0.00 < 0.05)
                        Performance has improved.

merge sorted utf8 dictionary
                        time:   [5.6368 ms 5.6413 ms 5.6461 ms]
                        change: [-14.343% -14.257% -14.164%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

sort merge utf8 dictionary
                        time:   [5.7874 ms 5.7961 ms 5.8055 ms]
                        change: [-14.409% -14.234% -14.047%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

merge sorted utf8 dictionary tuple
                        time:   [8.3309 ms 8.3364 ms 8.3423 ms]
                        change: [-10.213% -10.122% -10.033%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

sort merge utf8 dictionary tuple
                        time:   [9.6855 ms 9.7650 ms 9.8424 ms]
                        change: [-8.8845% -7.8176% -6.7832%] (p = 0.00 < 0.05)
                        Performance has improved.

merge sorted mixed dictionary tuple
                        time:   [14.514 ms 14.533 ms 14.553 ms]
                        change: [-5.1546% -5.0148% -4.8671%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

sort merge mixed dictionary tuple
                        time:   [15.994 ms 16.073 ms 16.151 ms]
                        change: [-4.7746% -4.1442% -3.4971%] (p = 0.00 < 0.05)
                        Performance has improved.

merge sorted mixed tuple
                        time:   [16.649 ms 16.675 ms 16.702 ms]
                        change: [-4.6957% -4.5328% -4.3548%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

sort merge mixed tuple  time:   [18.427 ms 18.477 ms 18.527 ms]
                        change: [-5.4182% -5.0185% -4.6040%] (p = 0.00 < 0.05)
                        Performance has improved.
```




# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

# Are there any user-facing changes?

Yes SortKeyCursor is removed, it is highly unlikely anyone was using this as it is very coupled with SortPreservingMerge

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->